pyproject.toml
[project]
name = "hcm-insight"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.18",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pandas>=2.2.3",
    "pydantic>=2.11.3",
    "pysqlite3>=0.5.4",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
    "secure>=1.0.1",
    "tenacity>=9.1.2",
    "tqdm>=4.67.1",
    "uvicorn>=0.34.2",
    "minio",
]

[project.optional-dependencies]
dev = [
    "coverage>=7.6.12",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
]

[tool.coverage.report]
omit = [
    "*/db_update.py",
    "*/llm_engine.py"
    ]
